Revert speak same toSpeak content in different hotbar slots
Because we can ONLY use the object id of the value of this.currentStack to determine whether the slot has changed. But sadly this class variable will also change when opening chests, opening craft table, use tools to break one block. We don't want to speak held item in these cases.

close #150